+++ plone/app/contenttypes/tests/test_collection.py
@@ -301,7 +301,7 @@ class PloneAppCollectionViewsIntegrationTest(unittest.TestCase):
         imagecount = wrapped.getFoldersAndImages()['total_number_of_images']
         # The current implementation for getFoldersAndImages will return
         # another_image under subfolder and also under folder
-        self.assertTrue(imagecount == 3)
+        self.assertEqual(imagecount, 3)
 
     def test_getFoldersAndImages_returning_images(self):
         portal = self.layer['portal']
@@ -335,7 +335,7 @@ class PloneAppCollectionViewsIntegrationTest(unittest.TestCase):
         wrapped = ICollection_behavior(collection)
         wrapped.query = query
         imagecount = wrapped.getFoldersAndImages()['total_number_of_images']
-        self.assertTrue(imagecount == 2)
+        self.assertEqual(imagecount, 2)
 
 
 class PloneAppCollectionEditViewsIntegrationTest(unittest.TestCase):