+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 1.1 (unreleased)
 ----------------
 
+- Collection: get ``querybuilderresults`` view instead of using the
+  ``QueryBuilder`` class directly.
+  [maurits]
+
+- Fix migration restoreReferencesOrder removes references
+  [joka]
+
 - Enable summary_view and all_content views for content types that
   have the collection behavior enabled.  Define collection_view for
   those types so you can view the results.  These simply show the